+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 ## [Unreleased]
 ### Added
 - make it possible to log documents fields in the status filter
+- implement a document reader
+- implement a document writer
 
 ### Changed